fix(build): clear every Error Prone warning on the compile path
The build annotated ten warnings per run, which is GitHub's per-check cap
rather than the real count — there were 28. Fixing only the visible ten
would have surfaced the next ten on the following run, so this clears all
of them.

Genuine defects:

- Locale-sensitive case folding on attack-type keywords, which matches
  differently under a Turkish locale.
- Objects.equals on long and int keys in the v9/IPFIX session tables and
  the TCP session, boxing both operands on a path that runs per data
  record.
- String.split for procfs columns and SSE query parameters: a regex
  recompiled per call that also drops trailing empty fields. Guava
  splitters state the intent instead.
- long to double widening in the Prometheus exposition, now an explicit
  cast so the conversion is deliberate rather than incidental.

Modifiers and comments: public members of private nested classes, and
comments marked /** on declarations Javadoc never renders — methods of
local classes, and one element that had ended up with two Javadoc blocks.

Switches: the MCP dispatch switch becomes an expression, which also gives
each case its own scope rather than sharing one across nine branches, and
the IPFIX selector-algorithm switch follows.

Three are suppressed with the reason recorded rather than changed:

- The parser Record types clash with java.lang.Record by name only. RFC
  3954 and RFC 7011 both call this a Record and nothing here refers to
  the JDK type; renaming would cost the protocol's vocabulary across
  every parser to settle a clash that cannot arise.
- The two packet-decode switches cannot become arrow switches: a case
  assigns a local declared before the switch and read after it, which
  arrow-case scoping breaks. Confirmed by trying it.
- Teardown's identity comparison is deliberate — addSuppressed rejects
  the throwable it is called on, and two equal-but-distinct throwables
  are two real failures.

Closes #483

+++ src/main/java/org/riptide/management/PrometheusExposition.java
@@ -53,14 +53,14 @@ final class PrometheusExposition {
         for (final Map.Entry<String, Counter> entry : registry.getCounters().entrySet()) {
             final String name = sanitize(entry.getKey());
             type(out, name, "counter");
-            sample(out, name, entry.getValue().getCount());
+            sample(out, name, (double) entry.getValue().getCount());
         }
 
         for (final Map.Entry<String, Meter> entry : registry.getMeters().entrySet()) {
             final String name = sanitize(entry.getKey());
             final Meter meter = entry.getValue();
             type(out, name, "counter");
-            sample(out, name, meter.getCount());
+            sample(out, name, (double) meter.getCount());
             // Dropwizard's own moving averages. Prometheus would normally derive a rate from the
             // counter, but exporting these costs nothing and they are what riptide's existing
             // meters were created to show.
@@ -75,7 +75,7 @@ final class PrometheusExposition {
             final Histogram histogram = entry.getValue();
             type(out, name, "summary");
             quantiles(out, name, histogram.getSnapshot(), 1.0d);
-            sample(out, name + "_count", histogram.getCount());
+            sample(out, name + "_count", (double) histogram.getCount());
         }
 
         for (final Map.Entry<String, Timer> entry : registry.getTimers().entrySet()) {
@@ -83,7 +83,7 @@ final class PrometheusExposition {
             final Timer timer = entry.getValue();
             type(out, name, "summary");
             quantiles(out, name, timer.getSnapshot(), 1.0d / TimeUnit.SECONDS.toNanos(1L));
-            sample(out, name + "_count", timer.getCount());
+            sample(out, name + "_count", (double) timer.getCount());
         }
 
         return out.toString();
